+++ src/controllers/taskController.js
@@ -1,10 +1,17 @@
 const { StatusCodes } = require('http-status-codes');
-const Task = require('../models/Task')
+const List = require('../models/List');
+const Task = require('../models/Task');
 
 const createTask = async (req, res) => {
+    const userId = req.userId;
     const { content, listId } = req.body;
 
     try {
+        const isOwner = await List.findListByOwner(userId, listId);
+        if (!isOwner) {
+            return res.status(StatusCodes.NOT_FOUND).end();
+        }
+
         const task = await Task.createTask(content, listId);
         return res.status(StatusCodes.CREATED).json(task);
     } catch (error) {
@@ -13,12 +20,16 @@ const createTask = async (req, res) => {
 }
 
 const updateTask = async (req, res) => {
+    const userId = req.userId;
     const { id } = req.params;
     const { content, done } = req.body;
 
     let updateRows = 0;
     try {
-        updateRows = await Task.updateTaskById(id, content, done);
+        const isOwner = await Task.findTaskByOwner(userId, id);
+        if (isOwner) {
+            updateRows = await Task.updateTaskById(id, content, done);
+        }
     } catch (error) {
         return res.status(StatusCodes.INTERNAL_SERVER_ERROR).end();
     }
@@ -31,11 +42,15 @@ const updateTask = async (req, res) => {
 
 
 const deleteTask = async (req, res) => {
+    const userId = req.userId;
     const { id } = req.params;
 
     let deleteRows = 0;
     try {
-        deleteRows = await Task.deleteTaskById(id);
+        const isOwner = await Task.findTaskByOwner(userId, id);
+        if (isOwner) {
+            deleteRows = await Task.deleteTaskById(id);
+        }
     } catch (error) {
         return res.status(StatusCodes.INTERNAL_SERVER_ERROR).end();
     }
@@ -47,8 +62,14 @@ const deleteTask = async (req, res) => {
 }
 
 const getAllTask = async (req, res) => {
+    const userId = req.userId;
     const { listId } = req.params;
     try {
+        const isOwner = await List.findListByOwner(userId, listId);
+        if (!isOwner) {
+            return res.status(StatusCodes.NOT_FOUND).end();
+        }
+
         const tasks = await Task.findAllTasks(listId);
         return res.status(StatusCodes.OK).json(tasks);
     } catch (error) {
@@ -57,8 +78,14 @@ const getAllTask = async (req, res) => {
 }
 
 const createBulkTask = async (req, res) => {
+    const userId = req.userId;
     const { listId, tasks } = req.body;
     try {
+        const isOwner = await List.findListByOwner(userId, listId);
+        if (!isOwner) {
+            return res.status(StatusCodes.NOT_FOUND).end();
+        }
+
         const task = await Task.createBulkTask(listId, tasks);
         return res.status(StatusCodes.CREATED).json(task);
     } catch (error) {
